TODO: fix youtube embed.
youtube embed is pain XD
TODO: fix that later. moving on to the rest of the site

Added template links for possible youutbe video's
Added extra comments where needed.

Page selection working
User can now go to the next page of the popular movies, and back the the previous page